requirements.txt
Django==4.0.6
python-dotenv==0.20.0
requests==2.28.1
pillow